+++ src/components/pages/Admin/components/RegistrationRequests/index.tsx
@@ -33,7 +33,6 @@ import {
   useDeclineChecklistMutation,
   useFetchCompanySearchQuery,
   useUpdateBPNMutation,
-  type ProgressButtonsType,
 } from 'features/admin/applicationRequestApiSlice'
 import { RequestList } from './components/RequestList'
 import { ServerResponseOverlay } from 'components/overlays/ServerResponse'
@@ -72,7 +71,6 @@ export default function RegistrationRequests() {
   const [successOverlay, setSuccessOverlay] = useState<boolean>(false)
   const [errorOverlay, setErrorOverlay] = useState<boolean>(false)
 
-  const [selectedButton, setSelectedButton] = useState<ProgressButtonsType>()
   const [statusConfirmationOverlay, setStatusConfirmationOverlay] =
     useState<boolean>(false)
   const [confirmCancelModalOpen, setConfirmCancelModalOpen] =
@@ -142,12 +140,6 @@ export default function RegistrationRequests() {
     setActionType('decline')
     setSelectedRequestId(id)
     setConfirmCancelModalOpen(true)
-  }
-
-  const onChipButtonSelect = (selected: ProgressButtonsType, id: string) => {
-    setSelectedButton(selected)
-    setSelectedRequestId(id)
-    setStatusConfirmationOverlay(true)
   }
 
   return (
@@ -225,7 +217,6 @@ export default function RegistrationRequests() {
           handleOverlayClose={() => {
             setStatusConfirmationOverlay(false)
           }}
-          selectedButton={selectedButton}
           selectedRequestId={selectedRequestId}
         />
       )}
@@ -273,9 +264,6 @@ export default function RegistrationRequests() {
           }}
           onConfirmationCancel={(id: string, name: string) => {
             onConfirmationCancel(id, name)
-          }}
-          onChipButtonSelect={(selected: ProgressButtonsType, id: string) => {
-            onChipButtonSelect(selected, id)
           }}
         />
       </div>